Dynamic patching of LocalStack's S3 pre-signed URL in dev
Pre-signed URLs returned from LocalStack's S3 have a host portion pointing to "localhost" (naturally).
However, this creates a problem whereby the expo dev builds on emulators and physical devices are unable to reach it (they should try to reach "10.0.2.2" or "192..." respectively).

Implement a patch in the axios intercepter which recursively patches the host portion of responses, such that "localhost:4567" (indicating pre-signed URLs) will point to the correct destination (10.0.2.2, or 192.168....etc...)